+++ build.gradle.kts
@@ -8,7 +8,6 @@ import org.cqfn.save.buildutils.disableRedundantKmpTargets
 import org.cqfn.save.buildutils.installGitHooks
 
 plugins {
-    kotlin("plugin.serialization") version Versions.kotlin apply false
     id("com.github.ben-manes.versions") version "0.39.0"
     id("com.cdsap.talaiot.plugin.base") version "1.4.2"
 }
