+++ save-plugins/warn-plugin/build.gradle.kts
@@ -1,46 +1,16 @@
 import org.cqfn.save.buildutils.configurePublishing
 
 plugins {
-    kotlin("multiplatform")
-    kotlin("plugin.serialization")
+    `kotlin-library`
 }
 
 kotlin {
-    jvm()
-    val nativeTargets = listOf(macosX64(), linuxX64(), mingwX64())
-
     sourceSets {
-        all {
-            languageSettings.useExperimentalAnnotation("kotlin.RequiresOptIn")
-            languageSettings.useExperimentalAnnotation("okio.ExperimentalFileSystem")
-        }
         val commonMain by getting {
             dependencies {
                 implementation(project(":save-common"))
                 implementation("org.jetbrains.kotlinx:kotlinx-serialization-core:${Versions.Kotlinx.serialization}")
             }
         }
-        val commonNonJsTest by creating {
-            dependencies {
-                implementation(kotlin("test-common"))
-                implementation(kotlin("test-annotations-common"))
-            }
-        }
-        val jvmTest by getting {
-            dependsOn(commonNonJsTest)
-            dependencies {
-                implementation(kotlin("test-junit5"))
-                implementation("org.junit.jupiter:junit-jupiter-engine:${Versions.junit}")
-            }
-        }
-        nativeTargets.forEach {
-            getByName("${it.name}Test").dependsOn(commonNonJsTest)
-        }
     }
-}
-
-configurePublishing()
-
-tasks.withType<org.jetbrains.kotlin.gradle.targets.jvm.tasks.KotlinJvmTest> {
-    useJUnitPlatform()
 }
